fix(Components): Move related components of components page
