+++ messages/en.json
@@ -1,5 +1,6 @@
 {
     "common": {
+        "": "",
         "Active": "Active",
         "Action": "Action",
         "Actions": "Actions",
