+++ messages/ja.json
@@ -1,5 +1,6 @@
 {
   "common": {
+    "": "",
     "APPROVED": "承認済",
     "Action": "アクション",
     "Actions": "アクション",
