+++ messages/vi.json
@@ -1,5 +1,6 @@
 {
   "common": {
+    "": "",
     "APPROVED": "Đã chấp thuận",
     "Action": "Hành động",
     "Actions": "Hành động",
