+++ messages/zh.json
@@ -1,5 +1,6 @@
 {
   "common": {
+    "": "",
     "APPROVED": "报告已批准",
     "Action": "操作",
     "Actions": "操作",
