+++ src/app/[locale]/components/components.module.css
@@ -10,6 +10,35 @@
   License-Filename: LICENSE
 */
 
+.components tr th:last-child {
+  width: 6%;
+}
+
+.components tr th:nth-child(4) {
+  width: 13%;
+}
+
+.components tr th:nth-child(3) {
+  max-width: 30%;
+}
+
+.components tr th:first-child {
+  width: 12%;
+}
+
+.components tr td {
+  word-wrap: break-word;
+}
+
+.delete-btn {
+  color: gray;
+  font-size: 14px;
+}
+
+.delete-btn:hover {
+  cursor: pointer;
+}
+
 .label {
   font-weight: bold;
 }
