+++ src/app/[locale]/components/components/DeleteComponentDialog.tsx
