--- src/components/component/components-table/ComponentsTable.module.css
@@ -1,40 +1,0 @@
-/*
-  Copyright (C) TOSHIBA CORPORATION, 2023. Part of the SW360 Frontend Project.
-  Copyright (C) Toshiba Software Development (Vietnam) Co., Ltd., 2023. Part of the SW360 Frontend Project.
-
-  This program and the accompanying materials are made
-  available under the terms of the Eclipse Public License 2.0
-  which is available at https://www.eclipse.org/legal/epl-2.0/
-
-  SPDX-License-Identifier: EPL-2.0
-  License-Filename: LICENSE
-*/
-
-.components tr th:last-child {
-  width: 6%;
-}
-
-.components tr th:nth-child(4) {
-  width: 13%;
-}
-
-.components tr th:nth-child(3) {
-  max-width: 30%;
-}
-
-.components tr th:first-child {
-  width: 12%;
-}
-
-.components tr td {
-  word-wrap: break-word;
-}
-
-.delete-btn {
-  color: gray;
-  font-size: 14px;
-}
-
-.delete-btn:hover {
-  cursor: pointer;
-}
