+++ src/components/sw360/SideBar/SideBar.module.css
