+++ src/components/sw360/SideBar/SideBar.tsx
